feat: restructure DJing section; move components to new directories and implement FollowMyDjContent
